Add canvas/tenant CRD e2e suite; fix bugs it caught (labels, delete revisions, logging)
Add hack/test-canvas-kind.sh (make integration-canvas-kind; nightly e2e matrix):
a single insecure NiFi cluster exercises NiFiParameterContext, NiFiProcessGroup
(bound to the context), NiFiControllerService (create DISABLED -> Enabled ->
ENABLED), NiFiLabel, and NiFiOutputPort, asserting each against NiFi's REST API,
then deletes them all and verifies removal + finalizer clearing. These kinds
reconcile by driving the NiFi REST API, so unit tests (which mock the client)
never exercised the real contract. The suite immediately found three real bugs:

1. Labels never became ready. NiFi's LabelDTO models width/height as Double and
   returns them with a decimal (200.0); LabelComponent.Width/Height were int32,
   so every label-create response failed to unmarshal. Now float64
   (pkg/nifi/labels_test.go guards it).

2. Deleting a canvas/tenant component deadlocked its finalizer whenever the
   component's stored revision fell behind NiFi's current one: the delete sent
   the stale Status.Revision.Version and NiFi rejected it with HTTP 400 "not the
   most up-to-date revision". OutputPort, InputPort, Label, ProcessGroup, Funnel,
   Connection, and ParameterContext all had this; they now refetch the current
   revision before deleting (matching ControllerService/ReportingTask), with a
   nil-safe fallback. Regression test asserts the refetched revision is used.

3. The operator emitted no logs at all -- cmd/manager never called
   ctrl.SetLogger, so controller-runtime discarded everything ("log.SetLogger was
   never called"). Added the standard zap logger (flag-configurable). This is
   what surfaced bug 2.

All verified live on kind (6 phases green) and by unit tests.

+++ pkg/nifi/client.go
@@ -383,13 +383,15 @@ type LabelEntity struct {
 }
 
 type LabelComponent struct {
-	ID            string            `json:"id,omitempty"`
-	ParentGroupID string            `json:"parentGroupId,omitempty"`
-	Label         string            `json:"label,omitempty"`
-	Position      *Position         `json:"position,omitempty"`
-	Width         int32             `json:"width,omitempty"`
-	Height        int32             `json:"height,omitempty"`
-	Style         map[string]string `json:"style,omitempty"`
+	ID            string    `json:"id,omitempty"`
+	ParentGroupID string    `json:"parentGroupId,omitempty"`
+	Label         string    `json:"label,omitempty"`
+	Position      *Position `json:"position,omitempty"`
+	// Width and Height are float64 because NiFi's LabelDTO models them as Double and returns
+	// them with a decimal (e.g. 200.0); unmarshaling that into an integer field fails.
+	Width  float64           `json:"width,omitempty"`
+	Height float64           `json:"height,omitempty"`
+	Style  map[string]string `json:"style,omitempty"`
 }
 
 type PortEntity struct {
